+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix compileOnly dependencies managed by spring BOM plugin (without version) generation in pom
+
 ### 2.0.0 (2020-01-17)
 * (breaking) Drop java 7 support
 * (breaking) Require gradle 5 or above (stable publishing automatic enabling removed as 5.0 enables it by default)